requirements.txt
langchain
langchain-openai
langchain-core
langchain-community
python-dotenv
langserve
pypdf
pymupdf
bs4
chromadb
faiss-cpu
groq
cassio
beautifulsoup4  
langchain-groq
wikipedia
arxiv
langchainhub
streamlit
sentence-transformers
langchain-objectbox
pinecone-client
pinecone-text
pinecone-notebooks